requirements.txt
fastapi
uvicorn
python-multipart
langchain
langchain-groq
langchain-community
pymupdf
chromadb
sentence-transformers
pydantic
python-dotenv